Fixed compilation on macOS

+++ src/serialize.h
@@ -558,7 +558,7 @@ public:
                 break;
             }
             int32_t idx = last + offset;
-            if (idx >= size) {
+            if (idx >= (int32_t)size) {
                 throw std::ios_base::failure("out of bounds index");
             }
             if (last != -1 && idx <= last) {
